# docs/swagger.yaml
basePath: /api
definitions:
...
  model.TeacherApplicationCreateRequest:
    properties:
      experience:
        type: string
      motivation:
        type: string
    required:
    - experience
    - motivation
    type: object
  model.TeacherApplicationQuery:
    properties:
      endDate:
        description: 'Format: YYYY-MM-DD'
        type: string
      pageIndex:
        type: integer
      pageSize:
        type: integer
      startDate:
        description: 'Format: YYYY-MM-DD'
        type: string
      status:
        type: integer
      userId:
        type: integer
    type: object
...
info:
  contact:
    name: LTEdu Support
  description: LTEdu 在线教育平台后端接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: LTEdu API
  version: "1.0"
paths:
...
  /v1/admin/teacher-applications/approve:
    post:
      consumes:
      - application/json
      description: 通过指定的教师资格申请（仅管理员）
      parameters:
      - description: 申请ID和备注
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 审批通过教师申请
      tags:
      - 教师申请
  /v1/admin/teacher-applications/detail:
    post:
      consumes:
      - application/json
      description: 获取指定教师申请的详细信息（仅管理员）
      parameters:
      - description: 申请ID
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取教师申请详情
      tags:
      - 教师申请
  /v1/admin/teacher-applications/list:
    post:
      consumes:
      - application/json
      description: 获取所有教师申请列表（仅管理员）
      parameters:
      - description: 查询条件
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.TeacherApplicationQuery'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取教师申请列表
      tags:
      - 教师申请
  /v1/admin/teacher-applications/reject:
    post:
      consumes:
      - application/json
      description: 拒绝指定的教师资格申请（仅管理员）
      parameters:
      - description: 申请ID和拒绝理由
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 拒绝教师申请
      tags:
      - 教师申请
...
  /v1/user/teacher/application:
    post:
      consumes:
      - application/json
      description: 获取当前用户的教师资格申请详情
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取我的教师申请
      tags:
      - 教师申请
  /v1/user/teacher/apply:
    post:
      consumes:
      - application/json
      description: 提交教师资格申请
      parameters:
      - description: 申请信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.TeacherApplicationCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 申请成为教师
      tags:
      - 教师申请
...
securityDefinitions:
  BearerAuth:
    description: '输入 "Bearer {token}" 格式的 JWT，例如: Bearer eyJhbGci...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"